Emily/requirements.txt
chainlit
huggingface_hub
textblob
nltk
langchain
python-dotenv
